requirements.txt
python-telegram-bot[job-queue]>=21.0
requests
python-dotenv
renault-api
aiohttp
